+++ src/gui/accountsettings.cpp
@@ -34,6 +34,7 @@
 
 #include <QDebug>
 #include <QDesktopServices>
+#include <QDir>
 #include <QListWidgetItem>
 #include <QMessageBox>
 #include <QAction>
