+++ src/gui/folder.h
@@ -26,15 +26,8 @@
 
 #include <csync.h>
 
-#include <QDir>
-#include <QHash>
-#include <QSet>
 #include <QObject>
 #include <QStringList>
-
-#include <QDebug>
-#include <QTimer>
-#include <qelapsedtimer.h>
 
 class QThread;
 class QSettings;
@@ -181,8 +174,7 @@ public:
 
      // Used by the Socket API
      SyncJournalDb *journalDb() { return &_journal; }
-
-     bool estimateState(QString fn, csync_ftw_type_e t, SyncFileStatus* s);
+     SyncEngine &syncEngine() { return *_engine; }
 
      RequestEtagJob *etagJob() { return _requestEtagJob; }
      qint64 msecSinceLastSync() const { return _timeSinceLastSyncDone.elapsed(); }
@@ -262,7 +254,6 @@ private slots:
     void etagRetreived(const QString &);
     void etagRetreivedFromSyncEngine(const QString &);
 
-    void slotAboutToPropagate(SyncFileItemVector& );
     void slotThreadTreeWalkResult(const SyncFileItemVector& ); // after sync is done
 
     void slotEmitFinishedDelayed();
@@ -303,15 +294,6 @@ private:
     /// The number of requested follow-up syncs.
     /// Reset when no follow-up is requested.
     int           _consecutiveFollowUpSyncs;
-
-    // SocketAPI: Cache files and folders that had errors so that they can
-    // get a red ERROR icon.
-    QSet<QString>   _stateLastSyncItemsWithErrorNew; // gets moved to _stateLastSyncItemsWithError at end of sync
-    QSet<QString>   _stateLastSyncItemsWithError;
-
-    // SocketAPI: A folder is tained if we got a file watcher notification
-    // for it. It's displayed as EVAL.
-    QSet<QString>   _stateTaintedFolders;
 
     SyncJournalDb _journal;
